Fix Rope Test
